chore: reorg solution
