--- xUnitRevit/CmdAvailabilityViews.cs
@@ -1,22 +1,0 @@
-﻿using Autodesk.Revit.DB;
-using Autodesk.Revit.UI;
-
-namespace xUnitRevit
-{
-  internal class CmdAvailabilityViews : IExternalCommandAvailability
-  {
-
-    /// <summary>
-    /// Command Availability - Views
-    /// </summary>
-    /// <param name="applicationData"></param>
-    /// <param name="selectedCategories"></param>
-    /// <returns></returns>
-    public bool IsCommandAvailable(UIApplication applicationData, CategorySet selectedCategories)
-    {
-      //Can be run from any view/state
-      return true;
-
-    }
-  }
-}
