+++ xUnitRevitUtils/xUnitRevitUtils2022/xUnitRevitUtils2022.csproj
@@ -8,7 +8,9 @@
 
 
   <ItemGroup>
-    <PackageReference Include="Speckle.Revit.API" Version="2022.0.2.1" />
+    <PackageReference Include="Speckle.Revit.API" Version="2022.0.2.1" >
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
     <PackageReference Include="xunit" Version="2.4.2" />
   </ItemGroup>
