+++ xUnitRevitUtils/xUnitRevitUtils2023/xUnitRevitUtils2023.csproj
@@ -1,7 +1,7 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras/3.0.44">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
+    <TargetFramework>netstandard2.0</TargetFramework>
 	  <Description>xUnit runner utilities for Revit 2023</Description>
     <PackageId>xUnitRevitUtils.2023</PackageId>
   </PropertyGroup>
